разделил на компоненты
