+++ public/index.html
@@ -35,6 +35,6 @@
   </section>
 </footer>
 
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
