+++ src/components/button.js
@@ -1,0 +1,7 @@
+const createShowMoreButton = function () {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createShowMoreButton};
